requirements.txt
attrs==24.3.0
azure-functions==1.21.3
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
confluent-kafka==2.6.1
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
Deprecated==1.2.15
firebase-admin==6.1.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.7.0
Flask-WTF==1.2.2
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.49.1
premailer==3.10.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.9.4
rsa==4.9
sanitize-filename==1.2.0
selenium==4.10.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
yagmail==0.15.293
